+++ src/main/java/com/majin/bit/controller/HorseController.java
@@ -37,110 +37,9 @@ import com.majin.bit.util.HorseCrawling;
 @Controller
 public class HorseController {
 	
-	@Autowired
-	private HorseService horseService;
+
 
 	@Autowired
 	private TrainerService trainerService;
-	
-	@RequestMapping(value="/admin")
-    public String apitest(){
-        return "admin/APIInsert";
-    }
-	
-	/**
-	 * 경주마 크롤링 후 MYSQL DB에 저장
-	 * @param request
-	 * @param model
-	 * @param meet
-	 * @return
-	 * @throws IOException
-	 * @throws ParseException
-	 */
-	@RequestMapping(value = "/horseapi", method = {RequestMethod.GET, RequestMethod.POST})
-	public String horse(HttpServletRequest request, Model model, String meet) throws IOException, ParseException {
-		String inputLine;
-		String buffer = "";
-		
-		StringBuilder urlBuilder = new StringBuilder("http://apis.data.go.kr/B551015/API8/raceHorseInfo");
-		urlBuilder.append("?" + URLEncoder.encode("ServiceKey", "UTF-8") + "=YsfeAQ1K0KPH1fOqYRLcvfqOP2P6Mo2iQOiZSumF4bMSlfyWjdPg4NWPu7Y5ms%2Fql9n2oi4dQbNq2bISj%2Bi4Hg%3D%3D");
-		urlBuilder.append("&" + URLEncoder.encode("numOfRows","UTF-8") + "=" + URLEncoder.encode("30000", "UTF-8")); /*한 페이지 결과 수 25186 */
-		urlBuilder.append("&" + URLEncoder.encode("pageNo","UTF-8") + "=" + URLEncoder.encode("1", "UTF-8")); /*페이지번호*/
-		urlBuilder.append("&" + URLEncoder.encode("meet","UTF-8") + "=" + URLEncoder.encode(meet, "UTF-8")); /*시행경마장구분(1.서울,2.제주,3.부산)*/
-		
-		System.out.println(urlBuilder.toString());
-		
-		URL url = new URL(urlBuilder.toString());
-		HttpURLConnection urlConnection = (HttpURLConnection)url.openConnection();
-		
-		urlConnection.setDoInput(true);
-		urlConnection.setDoOutput(true);
-		urlConnection.setRequestMethod("GET");
-		urlConnection.setRequestProperty("Accept", "application/json");
-		System.out.println("Response code : " + urlConnection.getResponseCode());
-		
-		BufferedReader bufferedReader;
-		if(urlConnection.getResponseCode() >= 200 && urlConnection.getResponseCode() <= 300) {
-			bufferedReader = new BufferedReader(new InputStreamReader(urlConnection.getInputStream(), "utf-8"));
-		}else {
-			bufferedReader = new BufferedReader(new InputStreamReader(urlConnection.getErrorStream(), "utf-8"));
-		}
-		
-		while((inputLine = bufferedReader.readLine()) != null) {
-			buffer += inputLine.trim();
-		}
-		
-		bufferedReader.close();
-		urlConnection.disconnect();
-		
-		JSONParser jsonParser = new JSONParser();
-		JSONObject jsonObject = (JSONObject) jsonParser.parse(buffer);
-		
-		List<HorseDto> horseList = horseService.horseInsert(jsonObject, meet);
-		
-		
-		model.addAttribute("horseList", horseList);
-		
-		return "admin/APIInsert :: #API";
-				
-	}
-	/**
-	 * 경주마 상세 정보
-	 * @param hrNo
-	 * @param meet
-	 * @param model
-	 * @return
-	 */
-	@RequestMapping(value = "/search/horseDetail", method = RequestMethod.POST)
-	public String horseDetail(String hrNo, String meet, Model model) {
-		
-		HorseDto horseDto = new HorseDto();
-		horseDto.setHrNo(hrNo);
-		horseDto.setMeet(meet);
-
-		model.addAttribute("horseDetail", horseService.searchOneHorse(horseDto));
-		
-		return "horseDetail";
-	}
-	
-	/**
-	 * 경주마 페이징
-	 * @param model
-	 * @param pageNo
-	 * @param search
-	 * @return
-	 */
-	@RequestMapping(value = "/search/horseSearchPaging", method = RequestMethod.POST)
-	public String horseSearchPaging(Model model, @RequestParam(defaultValue = "1")int pageNo, String search) {
-
-		List<HorseDto> horseList = horseService.searchHorse(search);
-		Pagination horsePagination = new Pagination(horseList.size(), pageNo, search);
-		List<HorseDto> horsePaging = horseService.searchPagingHorse(horsePagination);
-
-		model.addAttribute("searchHorse", horsePaging);
-		model.addAttribute("horsePagination", horsePagination);
-
-		return "horsePaging";
-	}
 
 }
