+++ src/main/java/com/majin/bit/controller/TrainerController.java
@@ -31,102 +31,9 @@ import com.majin.bit.util.TrainerCrawling;
 @Controller
 public class TrainerController {
 
-	@Autowired
-	private TrainerService trainerService;
-	
-	/**
-	 * 조교 크롤링 후 MYSQL DB에 저장
-	 * @param request
-	 * @param model
-	 * @param meet
-	 * @return
-	 * @throws IOException
-	 * @throws ParseException
-	 */
-	@RequestMapping(value = "/trapi", method = {RequestMethod.GET, RequestMethod.POST})
-	public String tr(HttpServletRequest request, Model model, String meet) throws IOException, ParseException {
-		String inputLine;
-		String buffer = "";
-	
-		StringBuilder urlBuilder = new StringBuilder("http://apis.data.go.kr/B551015/API19/trainerInfo");
-		urlBuilder.append("?" + URLEncoder.encode("ServiceKey", "UTF-8") + "=YsfeAQ1K0KPH1fOqYRLcvfqOP2P6Mo2iQOiZSumF4bMSlfyWjdPg4NWPu7Y5ms%2Fql9n2oi4dQbNq2bISj%2Bi4Hg%3D%3D");
-		urlBuilder.append("&" + URLEncoder.encode("numOfRows","UTF-8") + "=" + URLEncoder.encode("1000", "UTF-8")); /*한 페이지 결과 수*/
-		urlBuilder.append("&" + URLEncoder.encode("pageNo","UTF-8") + "=" + URLEncoder.encode("1", "UTF-8")); /*페이지번호*/
-		urlBuilder.append("&" + URLEncoder.encode("meet","UTF-8") + "=" + URLEncoder.encode(meet, "UTF-8")); /*시행경마장구분(1.서울,2.제주,3.부산)*/
 
-		System.out.println(urlBuilder.toString());
-		
-		URL url = new URL(urlBuilder.toString());
-		HttpURLConnection urlConnection = (HttpURLConnection)url.openConnection();
-		
-		urlConnection.setDoInput(true);
-		urlConnection.setDoOutput(true);
-		urlConnection.setRequestMethod("GET");
-		urlConnection.setRequestProperty("Accept", "application/json");
-		System.out.println("Response code : " + urlConnection.getResponseCode());
-		
-		BufferedReader bufferedReader;
-		if(urlConnection.getResponseCode() >= 200 && urlConnection.getResponseCode() <= 300) {
-			bufferedReader = new BufferedReader(new InputStreamReader(urlConnection.getInputStream(), "utf-8"));
-		}else {
-			bufferedReader = new BufferedReader(new InputStreamReader(urlConnection.getErrorStream(), "utf-8"));
-		}
-		
-		while((inputLine = bufferedReader.readLine()) != null) {
-			buffer += inputLine.trim();
-		}
-		
-		bufferedReader.close();
-		urlConnection.disconnect();
-		
-		JSONParser jsonParser = new JSONParser();
-		JSONObject jsonObject = (JSONObject) jsonParser.parse(buffer);
-		
-		List<TrDto> trList = trainerService.trainerInsert(jsonObject, meet);
-		
-		
-		model.addAttribute("trList", trList);
-		return "admin/APIInsert :: #API";
-		
-	}
 	
-	/**
-	 * 조교 상세 정보
-	 * @param trNo
-	 * @param meet
-	 * @param model
-	 * @return
-	 */
-	@RequestMapping(value = "/search/trainerDetail", method = RequestMethod.POST)
-	public String trainerDetail(String trNo, String meet, Model model) {
-		
-		TrDto trDto = new TrDto();
-		trDto.setTrNo(trNo);
-		trDto.setMeet(meet);
-		
-		model.addAttribute("trainerDetail",trainerService.searchOneTrainer(trDto));
-		
-		return "trainerDetail";
-	}
-	
-	/**
-	 * 조교 페이징
-	 * @param model
-	 * @param pageNo
-	 * @param search
-	 * @return
-	 */
-	@RequestMapping(value = "/search/trainerSearchPaging", method = RequestMethod.POST)
-	public String horseSearchPaging(Model model, int pageNo, String search) {
-		
-		List<TrDto> trainerList = trainerService.searchTrainer(search);
-		Pagination trainerPagination = new Pagination(trainerList.size(), pageNo, search);
-		List<TrDto> trainerPaging = trainerService.searchPagingTrainer(trainerPagination);
 
-		model.addAttribute("searchTrainer", trainerPaging);
-		model.addAttribute("trainerPagination", trainerPagination);
 
-		return "trainerPaging";
-	}
 	
 }
