+++ src/main/resources/templates/MultiCheck.html
@@ -551,7 +551,7 @@
 								<option value="01">01</option>
 
 						</select></td>
-						<td><button class="btn btn-light" th:onclick="multiSearchBtnStart()" type="button">전송</button></td>
+						<td><button class="btn btn-light" th:onclick="multiSearchBtnStart()" type="button">검색</button></td>
 					</tr>
 				</tbody>
 			</table>
