+++ src/main/resources/templates/admin/APIInsert.html
@@ -155,47 +155,47 @@
 		
 		let url;
 		$("#horseapi_s").click(function() {
-			url = "/horseapi";
+			url = "/admin/horseapi";
 			meet = {"meet" : "1"}
 			api(url, meet)
 		});
 		$("#horseapi_b").click(function() {
-			url = "/horseapi";
+			url = "/admin/horseapi";
 			meet = {"meet" : "3"}
 			api(url, meet)
 		});
 		$("#horseapi_j").click(function() {
-			url = "/horseapi";
+			url = "/admin/horseapi";
 			meet = {"meet" : "2"}
 			api(url, meet)
 		});
 		$("#jkapi_s").click(function() {
-			url = "/jkapi";
+			url = "/admin/jkapi";
 			meet = {"meet" : "1"}
 			api(url, meet)
 		});
 		$("#jkapi_b").click(function() {
-			url = "/jkapi";
+			url = "/admin/jkapi";
 			meet = {"meet" : "3"}
 			api(url, meet)
 		});
 		$("#jkapi_j").click(function() {
-			url = "/jkapi";
+			url = "/admin/jkapi";
 			meet = {"meet" : "2"}
 			api(url, meet)
 		});
 		$("#trapi_s").click(function() {
-			url = "/trapi";
+			url = "/admin/trapi";
 			meet = {"meet" : "1"}
 			api(url, meet)
 		});
 		$("#trapi_b").click(function() {
-			url = "/trapi";
+			url = "/admin/trapi";
 			meet = {"meet" : "3"}
 			api(url, meet)
 		});
 		$("#trapi_j").click(function() {
-			url = "/trapi";
+			url = "/admin/trapi";
 			meet = {"meet" : "2"}
 			api(url, meet)
 		});
