+++ src/main/resources/templates/trainerDetail.html
@@ -5,26 +5,22 @@
 
 <div id="information" class="container">
 	<h1 class="my-4">조교사 프로필</h1>
-	<table border=1>
-		<tr>
-			<th>성명</th>
-			<td>[[${trainerDetail.trName}]]</td>
-			<th>소속조</th>
-			<td>[[${trainerDetail.part}]]</td>
-			<th>데뷔날짜</th>
-			<td>[[${trainerDetail.debut}]]</td>
-			<th>최근 1년 전적</th>
-			<td>[[${trainerDetail.tYearTotal}]]</td>
-		</tr>
-		<tr>
-			<th>통산 전적</th>
-			<td>[[${trainerDetail.tAllTotal}]]</td>
-			<th>승률</th>
-			<td>[[${trainerDetail.tWinningP}]]</td>
-			<th>복승률</th>
-			<td>[[${trainerDetail.tComplementRyRate}]]</td>
-			<th>연승률</th>
-			<td>[[${trainerDetail.tConsecutiveWinningP}]]</td>
-		</tr>
-	</table>
+		<div class="row justify-content-md-center">
+		<div class="col-sm-1-top">성명</div>
+		<div class="col-sm-2" th:text="${trainerDetail.trName}"></div>
+		<div class="col-sm-1-top">소속조</div>
+		<div class="col-sm-1" th:text="${trainerDetail.part}"></div>
+		<div class="col-sm-1-top">데뷔날짜</div>
+		<div class="col-sm-1" th:text="${trainerDetail.debut}"></div>
+		<div class="col-sm-1-top">최근1년</div>
+		<div class="col-md-4" th:text="${trainerDetail.tYearTotal}"></div>
+		<div class="col-sm-1-top">승률</div>
+		<div class="col-sm-2" th:text="${trainerDetail.tWinningP}"></div>
+		<div class="col-sm-1-top">복승률</div>
+		<div class="col-sm-1" th:text="${trainerDetail.tComplementRyRate}"></div>
+		<div class="col-sm-1-top">연승률</div>
+		<div class="col-sm-1" th:text="${trainerDetail.tConsecutiveWinningP}"></div>
+		<div class="col-sm-1-top">통산전적 </div>
+		<div class="col-md" th:text="${trainerDetail.tAllTotal}"></div>
+	</div>
 </div>
